feat: User profile + settings menu in top right (AppHeader) instead of top/bottom left (AppSidebar)
