+++ frontend/components/Layout/LayoutParts/AppHeader.vue
@@ -58,17 +58,37 @@
       >
         <v-icon> {{ $globals.icons.search }}</v-icon>
       </v-btn>
-      <v-btn
-        v-if="loggedIn"
-        :variant="smAndUp ? 'text' : undefined"
-        :icon="xs"
-        @click="logout()"
-      >
-        <v-icon :start="smAndUp">
-          {{ $globals.icons.logout }}
-        </v-icon>
-        {{ smAndUp ? $t("user.logout") : "" }}
-      </v-btn>
+      <div v-if="loggedIn && sessionUser" class="mx-1">
+        <LanguageDialog v-model="languageDialog" />
+        <v-menu z-index="2020">
+          <template #activator="{ props }">
+            <v-btn v-bind="props" icon data-testid="user-menu-button">
+              <UserAvatar list :user-id="sessionUser.id" :tooltip="false" />
+            </v-btn>
+          </template>
+          <v-list density="comfortable" color="primary" data-testid="user-menu">
+            <v-list-item :title="sessionUser.fullName ?? undefined" :to="userProfileLink">
+              <template #prepend>
+                <!-- TODO: horizontally center avatar with other icons in menu (why isn't it automatically centered?) -->
+                <UserAvatar list :user-id="sessionUser.id" :tooltip="false" />
+              </template>
+            </v-list-item>
+            <v-list-item :prepend-icon="$globals.icons.heart" :title="$t('user.favorite-recipes')" :to="userFavoritesLink" />
+            <v-divider class="my-2" />
+            <v-list-item :prepend-icon="$globals.icons.translate" :title="$t('sidebar.language')" @click="languageDialog=true" />
+            <!-- TODO: prevent menu from closing when toggling light/dark mode -->
+            <v-list-item :prepend-icon="$vuetify.theme.current.dark ? $globals.icons.weatherSunny : $globals.icons.weatherNight" :title="$vuetify.theme.current.dark ? $t('settings.theme.light-mode') : $t('settings.theme.dark-mode')" @click="toggleDark" />
+            <!-- TODO: what was the use case for sidebar settings in logged-out state? -->
+            <v-divider v-if="loggedIn" class="my-2" />
+            <v-list-item v-if="loggedIn" :prepend-icon="$globals.icons.cog" :title="$t('profile.user-settings')" to="/user/profile" />
+            <v-list-item v-if="canManage" :prepend-icon="$globals.icons.manageData" :title="$t('data-pages.data-management')" to="/group/data" />
+            <v-divider v-if="isAdmin" class="my-2" />
+            <v-list-item v-if="isAdmin" :prepend-icon="$globals.icons.wrench" :title="$t('settings.admin-settings')" to="/admin/site-settings" />
+            <v-divider class="my-2" />
+            <v-list-item v-if="loggedIn" :prepend-icon="$globals.icons.logout" :title="$t('user.logout')" @click="logout()" />
+          </v-list>
+        </v-menu>
+      </div>
       <v-btn
         v-else
         variant="text"
@@ -87,9 +107,13 @@
 <script lang="ts">
 import { useLoggedInState } from "~/composables/use-logged-in-state";
 import RecipeDialogSearch from "~/components/Domain/Recipe/RecipeDialogSearch.vue";
+import UserAvatar from "~/components/Domain/User/UserAvatar.vue";
 
 export default defineNuxtComponent({
-  components: { RecipeDialogSearch },
+  components: {
+    RecipeDialogSearch,
+    UserAvatar,
+  },
   props: {
     menu: {
       type: Boolean,
@@ -99,12 +123,22 @@ export default defineNuxtComponent({
   setup() {
     const $auth = useMealieAuth();
     const { loggedIn } = useLoggedInState();
+    const isAdmin = computed(() => $auth.user.value?.admin);
+    const canManage = computed(() => $auth.user.value?.canManage);
+    const userFavoritesLink = computed(() => $auth.user.value ? `/user/${$auth.user.value.id}/favorites` : undefined);
+    const userProfileLink = computed(() => $auth.user.value ? "/user/profile" : undefined);
     const route = useRoute();
     const groupSlug = computed(() => route.params.groupSlug as string || $auth.user.value?.groupSlug || "");
     const { xs, smAndUp } = useDisplay();
 
     const routerLink = computed(() => groupSlug.value ? `/g/${groupSlug.value}` : "/");
     const domSearchDialog = ref<InstanceType<typeof RecipeDialogSearch> | null>(null);
+
+    const toggleDark = useToggleDarkMode();
+
+    const state = reactive({
+      languageDialog: false,
+    });
 
     function activateSearch() {
       domSearchDialog.value?.open();
@@ -136,18 +170,26 @@ export default defineNuxtComponent({
     }
 
     return {
+      ...toRefs(state),
       activateSearch,
+      canManage,
       domSearchDialog,
-      routerLink,
+      isAdmin,
       loggedIn,
       logout,
-      xs, smAndUp,
+      routerLink,
+      sessionUser: $auth.user,
+      smAndUp,
+      toggleDark,
+      userFavoritesLink,
+      userProfileLink,
+      xs,
     };
   },
 });
 </script>
 
-<style scoped>
+<style scoped lang="scss">
 .v-toolbar {
   z-index: 2010 !important;
 }
